e2e: move funs of framework/service to e2e/network
