+++ test/e2e/network/BUILD
@@ -17,6 +17,7 @@ go_library(
         "endpointslice.go",
         "example_cluster_dns.go",
         "firewall.go",
+        "fixture.go",
         "framework.go",
         "ingress.go",
         "ingress_scale.go",
@@ -29,6 +30,7 @@ go_library(
         "proxy.go",
         "service.go",
         "service_latency.go",
+        "util.go",
         "util_iperf.go",
     ],
     importpath = "k8s.io/kubernetes/test/e2e/network",
@@ -57,6 +59,7 @@ go_library(
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
         "//staging/src/k8s.io/client-go/tools/cache:go_default_library",
         "//staging/src/k8s.io/client-go/util/flowcontrol:go_default_library",
+        "//staging/src/k8s.io/client-go/util/retry:go_default_library",
         "//staging/src/k8s.io/client-go/util/workqueue:go_default_library",
         "//staging/src/k8s.io/cloud-provider:go_default_library",
         "//staging/src/k8s.io/legacy-cloud-providers/gce:go_default_library",
